added file checks for prev_delete.php and draft_delete.php

+++ scripts/prev_delete.php
@@ -5,7 +5,7 @@ require_once(DOKU_INC.'inc/io.php');
 
 $cname = $_REQUEST['draft_id'];
 $cname = urldecode($cname);
-
+if(!preg_match("#/data/cache/\w/[a-f0-9]{32}\.draft$#i", $cname)) return;
 if(file_exists($cname)) {
    if(unlink($cname)){
      echo  "$cname unlinked";
